refactor: 예외 처리 변경 (#114)
